+++ src/app/beer_garden/api/http/handlers/v1/command_publishing_blocklist.py
@@ -1,18 +1,5 @@
-from brewtils.models import Permissions
-
-import beer_garden.config as config
-from beer_garden.api.http.exceptions import BadRequest, NotFound, RoutingException
 from beer_garden.api.http.handlers import AuthorizationHandler
-from beer_garden.api.http.schemas.v1.command_publishing_blocklist import (
-    CommandPublishingBlocklistListInputSchema,
-    CommandPublishingBlocklistSchema,
-)
-from beer_garden.command_publishing_blocklist import (
-    command_publishing_blocklist_add,
-    command_publishing_blocklist_delete,
-)
-from beer_garden.db.mongo.models import CommandPublishingBlocklist, Garden
-from beer_garden.errors import RoutingRequestException
+from beer_garden.errors import EndpointRemovedException
 
 
 class CommandPublishingBlocklistPathAPI(AuthorizationHandler):
@@ -20,6 +7,7 @@ class CommandPublishingBlocklistPathAPI(AuthorizationHandler):
         """
         ---
         summary: Remove a command from event publishing block list
+        deprecated: true
         parameters:
           - name: command_publishing_id
             in: path
@@ -36,33 +24,11 @@ class CommandPublishingBlocklistPathAPI(AuthorizationHandler):
           50x:
             $ref: '#/definitions/50xError'
         tags:
-          - Command Block List
+          - Deprecated
         """
-
-        self.minimum_permission = Permissions.GARDEN_ADMIN.name
-        self.verify_user_global_permission()
-
-        blocked_command = CommandPublishingBlocklist.objects.get(
-            id=command_publishing_id
+        raise EndpointRemovedException(
+            message=("Command publishing blocklist API has been removed.")
         )
-
-        if blocked_command["namespace"] != config.get("garden.name"):
-            target_garden = Garden.objects.get(
-                namespaces__contains=blocked_command["namespace"],
-                connection_type__nin=[None, "LOCAL"],
-            )
-        else:
-            target_garden = Garden.objects.get(name=blocked_command["namespace"])
-        self.verify_user_permission_for_object(target_garden)
-
-        try:
-            command_publishing_blocklist_delete(blocked_command)
-        except RoutingRequestException:
-            raise RoutingException(
-                reason=f"Could not route to child garden: {blocked_command['namespace']}"
-            )
-
-        self.set_status(204)
 
 
 class CommandPublishingBlocklistAPI(AuthorizationHandler):
@@ -70,6 +36,7 @@ class CommandPublishingBlocklistAPI(AuthorizationHandler):
         """
         ---
         summary: Retrieve list of commands in publishing block list
+        deprecated: true
         responses:
           200:
             description: list of commands in publishing block list
@@ -80,28 +47,17 @@ class CommandPublishingBlocklistAPI(AuthorizationHandler):
           50x:
             $ref: '#/definitions/50xError'
         tags:
-          - Command Block List
+          - Deprecated
         """
-        self.minimum_permission = Permissions.GARDEN_ADMIN.name
-        self.verify_user_global_permission()
-
-        permitted_blocklist_entries = self.permissioned_queryset(
-            CommandPublishingBlocklist
+        raise EndpointRemovedException(
+            message=("Command publishing blocklist API has been removed.")
         )
-        response = {
-            "command_publishing_blocklist": CommandPublishingBlocklistSchema(many=True)
-            .dump(
-                CommandPublishingBlocklist.objects.filter(permitted_blocklist_entries)
-            )
-            .data
-        }
-
-        self.write(response)
 
     def post(self):
         """
         ---
         summary: Add a list of commands to event publishing block list
+        deprecated: true
         parameters:
           - name: CommandPublishingBlocklist
             in: body
@@ -120,51 +76,8 @@ class CommandPublishingBlocklistAPI(AuthorizationHandler):
           50x:
             $ref: '#/definitions/50xError'
         tags:
-          - Command Block List
+          - Deprecated
         """
-        self.minimum_permission = Permissions.GARDEN_ADMIN.name
-        self.verify_user_global_permission()
-
-        commands = self.schema_validated_body(CommandPublishingBlocklistListInputSchema)
-        checked_gardens = []
-
-        if len(commands["command_publishing_blocklist"]) == 0:
-            raise BadRequest(reason="Empty list was submitted")
-
-        for command in commands["command_publishing_blocklist"]:
-            if command["namespace"] not in checked_gardens:
-                try:
-                    if command["namespace"] != config.get("garden.name"):
-                        target_garden = Garden.objects.get(
-                            namespaces__contains=command["namespace"],
-                            connection_type__nin=[None, "LOCAL"],
-                        )
-                    else:
-                        target_garden = Garden.objects.get(name=command["namespace"])
-                    self.verify_user_permission_for_object(target_garden)
-                except NotFound:
-                    raise BadRequest(
-                        reason=f"Invalid garden name: {command['namespace']}"
-                    )
-                checked_gardens.append(command["namespace"])
-
-        added_commands = []
-        for command in commands["command_publishing_blocklist"]:
-            try:
-                blocked_command = command_publishing_blocklist_add(
-                    command, return_value=True
-                )
-                added_commands.append(blocked_command)
-            except RoutingRequestException:
-                raise RoutingException(
-                    reason=f"Could not route to child garden: {command['namespace']}"
-                )
-
-        response = {
-            "command_publishing_blocklist": CommandPublishingBlocklistSchema(many=True)
-            .dump(added_commands)
-            .data
-        }
-
-        self.set_status(201)
-        self.write(response)
+        raise EndpointRemovedException(
+            message=("Command publishing blocklist API has been removed.")
+        )