+++ src/app/test/db/mongo/models_test.py
@@ -15,7 +15,6 @@ import beer_garden.db.mongo.models
 from beer_garden.db.mongo.models import (
     Choices,
     Command,
-    CommandPublishingBlocklist,
     DateTrigger,
     Garden,
     Instance,
@@ -583,30 +582,6 @@ class TestUserToken:
         assert len(UserToken.objects.filter(id=user_token.id)) == 1
         User.objects.get(username="testuser").delete()
         assert len(UserToken.objects.filter(id=user_token.id)) == 0
-
-
-class TestCommandBlocklist:
-    namespace = "test"
-    system = "system_test"
-    command = "command_test"
-
-    @pytest.fixture()
-    def command_blocklist(self):
-        blocklist = CommandPublishingBlocklist(
-            namespace=self.namespace, system=self.system, command=self.command
-        ).save()
-
-        yield blocklist
-        blocklist.delete()
-
-    def test_blocklist_entries_are_required_to_be_unique(self, command_blocklist):
-        """Attempting to create a blocklist entry already in database should raise an
-        exception
-        """
-        with pytest.raises(NotUniqueError):
-            CommandPublishingBlocklist(
-                namespace=self.namespace, system=self.system, command=self.command
-            ).save()
 
 
 class TestGarden: