+++ src/app/test/publish_request_test.py
@@ -8,6 +8,7 @@ from brewtils.models import (
     Request,
     Subscriber,
     System,
+    Topic,
 )
 from mock import Mock
 
@@ -20,24 +21,24 @@ def command_topic_one():
 
 
 @pytest.fixture
-def command_topic_one_and_two():
-    return Command(name="command_one_topic", topics=["topic_1", "topic_2"])
+def command_topic_two():
+    return Command(name="command_two_topic", topics=["topic_1", "topic_2"])
 
 
 @pytest.fixture
 def command_topic_any():
-    return Command(name="command_one_topic", topics=["topic.*"])
+    return Command(name="command_any_topic", topics=["topic.*"])
 
 
 @pytest.fixture
-def localgarden_system(command_topic_one, command_topic_one_and_two, command_topic_any):
+def localgarden_system(command_topic_one, command_topic_two, command_topic_any):
     return System(
         name="localsystem",
         version="1.2.3",
         namespace="localgarden",
         local=True,
         instances=[Instance(name="default", status="RUNNING")],
-        commands=[command_topic_one, command_topic_one_and_two, command_topic_any],
+        commands=[command_topic_one, command_topic_two, command_topic_any],
     )
 
 
@@ -48,11 +49,70 @@ def localgarden(localgarden_system):
     )
 
 
+@pytest.fixture
+def topics():
+    return [
+        Topic(
+            name="topic_1",
+            subscribers=[
+                Subscriber(
+                    garden="localgarden",
+                    namespace="localgarden",
+                    system="localsystem",
+                    version="1.2.3",
+                    instance="default",
+                    command="command_one_topic",
+                    subscriber_type="DYNAMIC",
+                ),
+                Subscriber(
+                    garden="localgarden",
+                    namespace="localgarden",
+                    system="localsystem",
+                    version="1.2.3",
+                    instance="default",
+                    command="command_two_topic",
+                    subscriber_type="DYNAMIC",
+                ),
+            ],
+        ),
+        Topic(
+            name="topic_2",
+            subscribers=[
+                Subscriber(
+                    garden="localgarden",
+                    namespace="localgarden",
+                    system="localsystem",
+                    version="1.2.3",
+                    instance="default",
+                    command="command_two_topic",
+                    subscriber_type="DYNAMIC",
+                )
+            ],
+        ),
+        Topic(
+            name="topic.*",
+            subscribers=[
+                Subscriber(
+                    garden="localgarden",
+                    namespace="localgarden",
+                    system="localsystem",
+                    version="1.2.3",
+                    instance="default",
+                    command="command_any_topic",
+                    subscriber_type="DYNAMIC",
+                )
+            ],
+        ),
+    ]
+
+
 class TestSubscriptionEvent(object):
-    def test_newtopic(self, monkeypatch, localgarden):
+    def test_newtopic(self, monkeypatch, topics, localgarden):
         mock_process_request = Mock(return_value=None)
         monkeypatch.setattr(publish_request, "process_request", mock_process_request)
-        monkeypatch.setattr(publish_request, "get_gardens", Mock(return_value=[]))
+        monkeypatch.setattr(
+            publish_request, "get_all_topics", Mock(return_value=topics)
+        )
         monkeypatch.setattr(
             publish_request, "local_garden", Mock(return_value=localgarden)
         )
@@ -62,26 +122,18 @@ class TestSubscriptionEvent(object):
             metadata={"propagate": False, "topic": "newtopic"},
             payload=Request(),
         )
-        subscribers = [
-            Subscriber(
-                garden="localgarden",
-                system="localsystem",
-                version="1.2.3",
-                namespace="localgarden",
-                command="command_one_topic",
-                instance="default",
-            )
-        ]
-        publish_request.process_publish_event(
-            localgarden, event, subscribers=subscribers
-        )
 
-        assert mock_process_request.call_count == 3
+        publish_request.handle_event(event)
 
-    def test_topic_one(self, monkeypatch, localgarden):
+        assert mock_process_request.call_count == 0
+
+    def test_topic_one(self, monkeypatch, topics, localgarden):
         mock_process_request = Mock(return_value=None)
         monkeypatch.setattr(publish_request, "process_request", mock_process_request)
-        monkeypatch.setattr(publish_request, "get_gardens", Mock(return_value=[]))
+
+        monkeypatch.setattr(
+            publish_request, "get_all_topics", Mock(return_value=topics)
+        )
         monkeypatch.setattr(
             publish_request, "local_garden", Mock(return_value=localgarden)
         )
@@ -91,14 +143,16 @@ class TestSubscriptionEvent(object):
             metadata={"propagate": False, "topic": "topic_1"},
             payload=Request(),
         )
-        publish_request.process_publish_event(localgarden, event)
+        publish_request.handle_event(event)
 
         assert mock_process_request.call_count == 3
 
-    def test_topic_two(self, monkeypatch, localgarden):
+    def test_topic_two(self, monkeypatch, topics, localgarden):
         mock_process_request = Mock(return_value=None)
         monkeypatch.setattr(publish_request, "process_request", mock_process_request)
-        monkeypatch.setattr(publish_request, "get_gardens", Mock(return_value=[]))
+        monkeypatch.setattr(
+            publish_request, "get_all_topics", Mock(return_value=topics)
+        )
         monkeypatch.setattr(
             publish_request, "local_garden", Mock(return_value=localgarden)
         )
@@ -108,14 +162,16 @@ class TestSubscriptionEvent(object):
             metadata={"propagate": False, "topic": "topic_2"},
             payload=Request(),
         )
-        publish_request.process_publish_event(localgarden, event)
+        publish_request.handle_event(event)
 
         assert mock_process_request.call_count == 2
 
-    def test_topic_wildcard(self, monkeypatch, localgarden):
+    def test_topic_wildcard(self, monkeypatch, topics, localgarden):
         mock_process_request = Mock(return_value=None)
         monkeypatch.setattr(publish_request, "process_request", mock_process_request)
-        monkeypatch.setattr(publish_request, "get_gardens", Mock(return_value=[]))
+        monkeypatch.setattr(
+            publish_request, "get_all_topics", Mock(return_value=topics)
+        )
         monkeypatch.setattr(
             publish_request, "local_garden", Mock(return_value=localgarden)
         )
@@ -125,14 +181,16 @@ class TestSubscriptionEvent(object):
             metadata={"propagate": False, "topic": "topic_3"},
             payload=Request(),
         )
-        publish_request.process_publish_event(localgarden, event)
+        publish_request.handle_event(event)
 
         assert mock_process_request.call_count == 1
 
-    def test_topic_start_substring(self, monkeypatch, localgarden):
+    def test_topic_start_substring(self, monkeypatch, topics, localgarden):
         mock_process_request = Mock(return_value=None)
         monkeypatch.setattr(publish_request, "process_request", mock_process_request)
-        monkeypatch.setattr(publish_request, "get_gardens", Mock(return_value=[]))
+        monkeypatch.setattr(
+            publish_request, "get_all_topics", Mock(return_value=topics)
+        )
         monkeypatch.setattr(
             publish_request, "local_garden", Mock(return_value=localgarden)
         )
@@ -142,14 +200,16 @@ class TestSubscriptionEvent(object):
             metadata={"propagate": False, "topic": "topic"},
             payload=Request(),
         )
-        publish_request.process_publish_event(localgarden, event)
+        publish_request.handle_event(event)
 
         assert mock_process_request.call_count == 1
 
-    def test_topic_non_match(self, monkeypatch, localgarden):
+    def test_topic_non_match(self, monkeypatch, topics, localgarden):
         mock_process_request = Mock(return_value=None)
         monkeypatch.setattr(publish_request, "process_request", mock_process_request)
-        monkeypatch.setattr(publish_request, "get_gardens", Mock(return_value=[]))
+        monkeypatch.setattr(
+            publish_request, "get_all_topics", Mock(return_value=topics)
+        )
         monkeypatch.setattr(
             publish_request, "local_garden", Mock(return_value=localgarden)
         )
@@ -159,6 +219,6 @@ class TestSubscriptionEvent(object):
             metadata={"propagate": False, "topic": "badTopic"},
             payload=Request(),
         )
-        publish_request.process_publish_event(localgarden, event)
+        publish_request.handle_event(event)
 
         assert mock_process_request.call_count == 0